Move Solarium adapter to package
