+++ lib/Adapter/Solarium/composer.json
@@ -1,0 +1,23 @@
+{
+    "name": "pagerfanta/solarium-adapter",
+    "type": "library",
+    "description": "Pagerfanta adapter for Solarium",
+    "keywords": ["pagerfanta", "solarium"],
+    "license": "MIT",
+    "require": {
+        "php": "^7.2",
+        "pagerfanta/core": "^2.4",
+        "solarium/solarium": "^2.3 || ^3.0 || ^4.0 || ^5.0"
+    },
+    "autoload": {
+        "psr-4": {
+            "Pagerfanta\\Solarium\\": "./"
+        }
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-2.x": "2.x-dev"
+        }
+    },
+    "minimum-stability": "dev"
+}
